recorded tab
