+++ examples/chat-crx/public/styles/sidepanel.css
@@ -11,13 +11,17 @@ body {
     height: 100vh;
     display: flex;
     flex-direction: column;
+    margin: 0;
+    padding: 0;
+    overflow: hidden;
+    background: none;
 }
 
 .header {
-    background: rgba(255, 255, 255, 0.1);
+    background: none;
+    border-bottom: none;
     backdrop-filter: blur(10px);
     padding: 1rem;
-    border-bottom: 1px solid rgba(255, 255, 255, 0.2);
     display: flex;
     flex-direction: row;
     flex-wrap: wrap;
@@ -353,4 +357,86 @@ input:checked+.toggle-slider:before {
     .attach-button {
         flex-shrink: 0;
     }
+}
+
+/* Tab Container Styles */
+.tab-container {
+    display: flex;
+    flex-direction: column;
+    height: 100vh;
+    background: linear-gradient(135deg, #667eea 0%, #764ba2 100%);
+}
+
+.tab-header {
+    display: flex;
+    background: rgba(255, 255, 255, 0.1);
+    backdrop-filter: blur(10px);
+    padding: 1rem;
+    gap: 1rem;
+    border-bottom: 1px solid rgba(255, 255, 255, 0.2);
+}
+
+.tab-button {
+    background: transparent;
+    border: none;
+    color: rgba(255, 255, 255, 0.7);
+    padding: 0.5rem 1rem;
+    font-size: 1rem;
+    cursor: pointer;
+    transition: all 0.2s ease;
+    border-radius: 0.5rem;
+    font-weight: 500;
+}
+
+.tab-button:hover {
+    background: rgba(255, 255, 255, 0.1);
+    color: white;
+}
+
+.tab-button.active {
+    background: rgba(255, 255, 255, 0.2);
+    color: white;
+}
+
+.tab-content {
+    display: none;
+    flex: 1;
+    overflow: hidden;
+}
+
+.tab-content.active {
+    display: flex;
+    flex-direction: column;
+}
+
+/* Demos Tab Styles */
+.demos-container {
+    flex: 1;
+    padding: 1rem;
+    overflow-y: auto;
+}
+
+.demos-list {
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+}
+
+.empty-state {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    color: rgba(255, 255, 255, 0.7);
+    gap: 1rem;
+}
+
+.empty-state svg {
+    opacity: 0.7;
+}
+
+.empty-state p {
+    font-size: 1rem;
 }
